Replace image_write_to_array with image_write_to_binary While investigations are under way to determine the safest way to pass image data references to other consumers, add `vips_image_write_to_binary` to take the raw image data (the matrix) and convert it to a binary that can be safely passed around.
This will allow for development and testing of interoperability
with other libraries like eVision.

Performance isn't going to be great since there is a lot
of memory copying, but its enough to get started.

The return format is a direct copy of the `Nx.Tensor.t`
struct in an attempt to find a canonical interoperability
structure for sharing data.